requirements.txt
google-generativeai>=0.3.0
google-cloud-texttospeech>=2.14.0
moviepy>=1.0.3
Pillow>=10.0.0
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
requests>=2.31.0
python-dotenv>=1.0.0
PyYAML>=6.0
discord-webhook>=1.3.0
